Change the size of 'serverid' from smallint to int
git-svn-id: https://bzflag.svn.sourceforge.net/svnroot/bzflag/trunk@14339 08b3d480-bf2c-0410-a26f-811ee3361c24

+++ db/bzfls2/database.sql
@@ -3,7 +3,7 @@
 -- 
 
 CREATE TABLE `servers` (
-  `serverid` smallint(5) unsigned NOT NULL,
+  `serverid` int(10) unsigned NOT NULL,
   `name` varchar(64) NOT NULL,
   `port` smallint(5) unsigned NOT NULL,
   `ipaddress` varchar(15) NOT NULL,
